// inertia/types/pagination.ts
export type PaginationMeta = {
  total: number
  perPage: number
  currentPage: number
  lastPage: number
  firstPage: number
  firstPageUrl: string
  lastPageUrl: string
  nextPageUrl: string | null
  previousPageUrl: string | null
}

export type PaginatedData<T> = {
  data: T[]
  meta: PaginationMeta
}